Add handling for nil values
If the value in the database is null it
should also be null when inserting it again.
Before an empty string was generated and this
can lead to problems with indexes and constraints.

A combined unique index like locale,uri where the uri
is optional would throw an duplicate Error when you
insert multiple values like this:

'de_ch',''
'de_ch',''

The correct way to insert these:

'de_ch',null
'de_ch',null

And will not throw any errors.